physical-ai-and-humanoid-robotics-textbook/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.38.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
qdrant-client==1.12.1
cohere==5.13.4
openai-agents==0.6.3
python-dotenv==1.2.1
pydantic-settings==2.12.0
asyncpg==0.30.0
python-multipart==0.0.20
pydantic==2.12.5
aiofiles==24.1.0
python-frontmatter==1.1.0
tiktoken==0.9.0